pyproject.toml
[project]
name = "redact-pdf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pdf-redactor>=0.0.1",
    "pymupdf>=1.24.14",
    "python-dotenv>=1.0.1",
]

[tool.mypy]
disable_error_code = "import-untyped"
